add user cards, begin dev on game ui
